Had to revert the whole project.
It seemed a switch to Android in build settings and possibly upgrading the interaction toolkit corrupted the project. After some failed attempts at fixing the problems a revert was the least time consuming option.
Despite these problems i actually managed to build an apk and run it on the headset.